build scripts: update windows build scripts to handle freeglut

+++ Build/freeglut/clang_win/make_freeglut.bat
@@ -1,12 +1,10 @@
 @echo off
 
-:: setup compiler environment
-call ..\..\..\Utilities\Scripts\setup_compilers.bat clang
 
 ::*** clean old files
 echo *** removing old files
 git clean -dxf
-rmdir /s /q  ..\..\..\..\libs\freeglut
+if exist ..\..\..\..\libs\freeglut rmdir /s /q  ..\..\..\..\libs\freeglut
 mkdir ..\..\..\..\libs\freeglut
 
 ::*** configure
@@ -16,8 +14,8 @@ cmake ../../../../freeglut ^
   -DCMAKE_INSTALL_PREFIX=../../../../libs/freeglut ^
   -DCMAKE_BUILD_TYPE=Release ^
   -DCMAKE_C_COMPILER=clang-cl ^
+  -DCMAKE_LINKER=link ^
   -DCMAKE_C_FLAGS="/utf-8 /O2 /Gy" ^
-  -DCMAKE_CXX_FLAGS="/utf-8 /O2 /Gy" ^
   -DCMAKE_MSVC_RUNTIME_LIBRARY=MultiThreaded ^
   -DFREEGLUT_BUILD_SHARED_LIBS=OFF ^
   -DFREEGLUT_BUILD_STATIC_LIBS=ON ^
@@ -28,14 +26,3 @@ ninja
 ::*** install
 echo *** installing
 ninja install 
-
-
-::cmake ../../../../freeglut ^
-::  -G "Ninja" ^
-::  -DCMAKE_INSTALL_PREFIX=../../../../libs/freeglut ^
-::  -DCMAKE_BUILD_TYPE=Release ^
-::  -DCMAKE_C_COMPILER=clang-cl ^
-::  -DCMAKE_C_FLAGS="/utf-8" ^
-::  -DFREEGLUT_BUILD_SHARED_LIBS=OFF ^
-::  -DFREEGLUT_BUILD_STATIC_LIBS=ON ^
-::  -DFREEGLUT_BUILD_DEMOS=OFF
